+++ src/Components/DeleteCar/CarCard.jsx
@@ -2,7 +2,7 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import axios from 'axios';
-import classes from './DeleteCar.module.css';
+import styles from './DeleteCar.module.css';
 
 const DeleteCarUi = ({
   id, image, model, driver_name, cars, setCars,
@@ -10,7 +10,7 @@ const DeleteCarUi = ({
   const [errorMessage, setErrorMessage] = useState('');
   const removeCar = (id) => {
     axios
-      .delete(`http://localhost:3000/api/v1/car/${id}`)
+      .delete(`https://luxy-rides-api.herokuapp.com/api/v1/car/${id}`)
       .then(() => {
         let filteredCars = [];
         filteredCars = cars.filter((car) => car.id !== id);
@@ -23,10 +23,10 @@ const DeleteCarUi = ({
 
   return (
     <>
-      <p className={`${classes.errorMsg} fs-5`}>{errorMessage}</p>
-      <section className={`${classes.carCnt} mx-4`} key={id}>
-        <div className={classes.imgCnt}>
-          <img src={image} alt={model} className={classes.carImg} />
+      <p className={`${styles.errorMsg} fs-5`}>{errorMessage}</p>
+      <section className={`${styles.carCnt} mx-4`} key={id}>
+        <div className={styles.imgCnt}>
+          <img src={image} alt={model} className={styles.carImg} />
         </div>
         <div className="model-detail d-flex flex-column justify-content-center align-items-center">
           <h3 className="h4 mt-3">{model}</h3>
