+++ src/config/routes.js
@@ -1,13 +1,13 @@
-import CarDetails from '../Components/Cars/CarDetails';
-import Home from '../Components/Home/Home';
-import Login from '../Components/Login/Login';
 import PrivateRoutes from '../Components/PrivateRoutes/PrivateRoutes';
+import Splash from '../Components/SplashScreen/Splash';
+import Login from '../Components/Login/Login';
 import Register from '../Components/Register/Register';
-import AddCar from '../Components/Cars/AddCar';
+import Home from '../Components/Home/Home';
+import CarDetails from '../Components/CarDetails/CarDetails';
+import AddCar from '../Components/AddCar/AddCar';
+import DeleteCar from '../Components/DeleteCar/DeleteCar';
 import Reserve from '../Components/Reserve/Reserve';
 import MyReservations from '../Components/MyReservations/MyReservations';
-import Splash from '../Components/SplashScreen/Splash';
-import DeleteCar from '../Components/DeleteCar/DeleteCar';
 
 const routes = [
   {
